requirements.txt
# Core
fastapi>=0.109.0
uvicorn[standard]>=0.27.0
python-dotenv>=1.0.0

# Database
sqlalchemy>=2.0.0
asyncpg>=0.29.0
pgvector>=0.2.4

# Validation & XML
pydantic>=2.6.0
lxml>=5.1.0

# AI & Embeddings
sentence-transformers>=2.3.0
numpy>=1.26.0

# Groq (for future text generation)
groq>=0.4.0
